tools/barli/requirements.txt
rumps>=0.4.0
watchdog>=4.0.0
PyYAML>=6.0
